fix: move types to site-config
