+++ app/components/SectionCardCarousel.tsx
@@ -1,7 +1,7 @@
 import { Card, type CardProps } from "@teamimpact/veda-ui-blocks";
 import type { ReactNode } from "react";
 import { Section, SectionHeading, type SectionProps } from "@/app/components/";
-import type { IterableItemWithId } from "@/app/components/types";
+import type { IterableItemWithId } from "@/app/site-config/types";
 
 export type SectionCardCarouselProps = SectionProps & {
   sectionHeading?: ReactNode;
