+++ app/components/SectionCardSimple.tsx
@@ -2,7 +2,7 @@ import type { CardSimpleProps } from "@teamimpact/veda-ui-blocks";
 import { CardSimple } from "@teamimpact/veda-ui-blocks";
 import type { ReactNode } from "react";
 import { Section, SectionHeading, type SectionProps } from "@/app/components";
-import type { IterableItemWithId } from "@/app/components/types";
+import type { IterableItemWithId } from "@/app/site-config/types";
 
 export type SectionCardSimpleProps = SectionProps & {
   sectionHeading?: ReactNode;
