--- app/components/types.ts
@@ -1,1 +1,0 @@
-export type IterableItemWithId<T> = T & { id: string };
