+++ app/site-config/dataset/datasets.ts
@@ -1,4 +1,4 @@
-import type { IterableItemWithId } from "@/app/components/types";
+import type { IterableItemWithId } from "@/app/site-config/types";
 
 type Dataset = {
   contentType: "dataset";
