+++ app/site-config/types.ts
@@ -1,0 +1,3 @@
+export type IterableItemWithId<T> = T & { id: string };
+
+export type Theme = "respond" | "build" | "prepare" | "recover";
